+++ index.d.ts
@@ -443,7 +443,6 @@ declare namespace Moleculer {
 		fatal(message: string, err?: Error, needExit?: boolean): void;
 		loadServices(folder?: string, fileMask?: string): number;
 		loadService(filePath: string): Service;
-		hotReloadService(service: Service): Service;
 		createService(schema: ServiceSchema, schemaMods?: ServiceSchema): Service;
 		destroyService(service: Service): PromiseLike<void>;