+++ test/integration/broker.spec.js
@@ -332,3 +332,43 @@ describe("Test cachers", () => {
 
 });
 
+describe("Test async current Context store", () => {
+
+	let broker = new ServiceBroker({
+		logger: false
+	});
+
+	broker.createService({
+		name: "user",
+		actions: {
+			save() {
+				return this.Promise.resolve()
+					.then(() => this.saveUser());
+			}
+		},
+
+		methods: {
+			saveUser() {
+				const ctx = this.broker.getCurrentContext();
+				return Object.assign(ctx.params, { id: 5 });
+			}
+		}
+	});
+
+	beforeAll(() => broker.start());
+	afterAll(() => broker.stop());
+
+	it("should find context in method", () => {
+		const params = { name: "John", age: 33 };
+
+		return broker.call("user.save", params).catch(protectReject).then(res => {
+			expect(res).toEqual({
+				id: 5,
+				name: "John",
+				age: 33
+			});
+		});
+	});
+
+});
+